Fixed VS2005 projects. Removed warnings
git-svn-id: https://unittest-cpp.svn.sourceforge.net/svnroot/unittest-cpp@88 feadf2a4-a60d-0410-a25d-8df00a7d1702

+++ UnitTest++/src/MemoryOutStream.cpp
@@ -28,7 +28,7 @@ char const* MemoryOutStream::GetText() const
 
 MemoryOutStream& MemoryOutStream::operator << (char const* txt)
 {
-    int const bytesLeft = m_bufferSize - std::strlen(m_buffer);
+    size_t const bytesLeft = m_bufferSize - std::strlen(m_buffer);
     std::strncat(m_buffer, txt, bytesLeft-1);
     return *this;
 }
